Fix timer tick overflow (hopefully)

+++ pythonplugin.cpp
@@ -318,6 +318,9 @@ PLUGIN_EXPORT int PLUGIN_CALL AmxUnload( AMX *amx )
 		struct timeval tv;
 		if(gettimeofday(&tv, NULL) != 0) return 0;
 
-		return (tv.tv_sec * 1000) + (tv.tv_usec / 1000);
+		return (
+			((unsigned long long)tv.tv_sec * 1000ULL)
+			+ (unsigned long long)(tv.tv_usec / 1000)
+		);
 	}
 #endif
